requirements.txt
django==3.1.4
django-debug-toolbar==3.2
django-extensions==3.1.0
django-bootstrap4==2.3.1
pillow
django-taggit
django-taggit-templatetags
django-taggit-templatetags2
django-summernote
django-static-fontawesome
python-magic
django-environ
mysqlclient
django-sendgrid-v5
sendgrid
django-storages 
boto3